% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{v_caching}
\alias{v_caching}
\alias{v_cache_dir}
\title{Get or set fivbvis caching behaviour}
\usage{
v_caching(caching)

v_cache_dir(cache_dir)
}
\arguments{
\item{caching}{logical or string: if \code{TRUE}, used cached results; if \code{FALSE}, no caching; if \code{"refresh"}, refresh the cache}

\item{cache_dir}{string: either
\itemize{
\item "session" - use a session-specific directory for caching, meaning that when R is closed and restarted, the cache will be cleared and a new one used
\item "user" - use a persistent, user-specific directory (as returned by \code{\link[rappdirs]{user_data_dir}}
\item or a string giving a path to a specific directory to use. The directory must already exist
}}
}
\value{
The caching setting (for \code{v_caching}) or cache directory (for \code{v_cache_dir}) after applying any provided parameters
}
\description{
By default, fivbvis starts with session-specific caching (equivalent to running \verb{v_caching(TRUE); v_cache_dir("session")}).
}
\examples{

v_caching(TRUE) ## turn caching on
v_caching(FALSE) ## turn caching off
v_caching("refresh") ## contents in cache will be updated

v_cache_dir() ## return current cache dir
\dontrun{
  v_cache_dir("session") ## use per-session cache
  v_cache_dir("user") ## use persistent, user-specific cache
  v_cache_dir("c:/my/cache/dir") ## use specific dir
}

}
