% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{v_flag}
\alias{v_flag}
\title{Retrieve a country flag image}
\usage{
v_flag(country_code)
}
\arguments{
\item{country_code}{string: the ISO 3166-1 alpha-2 code of the country. For federations, you need to convert the federation code to the equivalent country code. If there is no equivalent country (for ENG, NIR, SCO, WAL), you can use the 3 letters federation code to retrieve the flag.}
}
\value{
Path to image file
}
\description{
Retrieve a country flag image
}
\examples{
\dontrun{
  v_flag("AU")
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#Country\%20flags\%20(small).html}
}
