% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beach.R
\name{v_get_beach_match}
\alias{v_get_beach_match}
\title{Get a beach volleyball match}
\usage{
v_get_beach_match(no, fields)
}
\arguments{
\item{no}{integer: match number}

\item{fields}{character: fields to return}
}
\value{
A data.frame
}
\description{
Get a beach volleyball match
}
\examples{
\dontrun{
  v_get_beach_match(15592)
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetBeachMatch.html}
}
