% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beach.R
\name{v_get_beach_round_ranking}
\alias{v_get_beach_round_ranking}
\title{Request to get the ranking of a beach volleyball round}
\usage{
v_get_beach_round_ranking(no, fields = v_fields("Beach Round Ranking"))
}
\arguments{
\item{no}{integer: number of the beach volleyball round}

\item{fields}{character: fields to return}
}
\value{
A data.frame
}
\description{
Request to get the ranking of a beach volleyball round
}
\examples{
\dontrun{
  v_get_beach_round_ranking(980)
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetBeachRoundRanking.html}
}
