% Generated by roxygen2: do not edit by hand

\name{v_get_beach_team}
\alias{v_get_beach_team}
\title{Get a beach volleyball team}
\usage{
v_get_beach_team(no, fields)
}
\arguments{
\item{no}{integer: team number}

\item{fields}{character: fields to return}
}
\value{
A data.frame
}
\description{
Get a beach volleyball team
}
\examples{
\dontrun{
  v_get_beach_team(375442)
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetBeachTeam.html}
}
