% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beach.R
\name{v_get_beach_tournament}
\alias{v_get_beach_tournament}
\title{Get a beach tournament}
\usage{
v_get_beach_tournament(no, fields)
}
\arguments{
\item{no}{integer: the number of the tournament}

\item{fields}{character: fields to return}
}
\value{
A data.frame
}
\description{
Get a beach tournament
}
\examples{
\dontrun{
  v_get_beach_tournament(502)
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetBeachTournament.html}
}
