% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beach.R
\name{v_get_beach_tournament_ranking}
\alias{v_get_beach_tournament_ranking}
\title{Request to get the ranking of a beach volleyball tournament}
\usage{
v_get_beach_tournament_ranking(
  no,
  phase,
  fields = v_fields("Beach Tournament Ranking")
)
}
\arguments{
\item{no}{integer: number of the beach volleyball tournament}

\item{phase}{string: phase for which to return the ranking ("Qualification" or "MainDraw")}

\item{fields}{character: fields to return}
}
\value{
A data.frame
}
\description{
Request to get the ranking of a beach volleyball tournament
}
\examples{
\dontrun{
  v_get_beach_tournament_ranking(503)
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetBeachTournamentRanking.html}
}
