% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beach.R
\name{v_get_beach_world_tour_ranking}
\alias{v_get_beach_world_tour_ranking}
\title{Request to get beach volleyball World Tour rankings}
\usage{
v_get_beach_world_tour_ranking(
  gender,
  number,
  referencedate,
  fields = v_fields("Beach World Tour Ranking")
)
}
\arguments{
\item{gender}{character: gender of the teams to retrieve the ranking for}

\item{number}{integer: number of entries to return}

\item{referencedate}{date: reference date of the ranking when the ranking has been calculated. YYY-MM-DD. if there is no World Tour ranking calculated at the specified date, the returned ranking will be empty. if this parameter is not specified, the web service will return the ranking for the latest reference date.}

\item{fields}{character: fields to return}
}
\value{
A data.frame
}
\description{
Request to get a beach volleyball Olympic Selection ranking
}
\examples{
\dontrun{
  v_get_beach_world_tour_ranking("W")
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetBeachWorldTourRanking.html}
}
