% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{v_get_event}
\alias{v_get_event}
\title{Get a volleyball event}
\usage{
v_get_event(no, fields)
}
\arguments{
\item{no}{integer: match number}

\item{fields}{character: fields to return}
}
\value{
A data.frame
}
\description{
Get a volleyball event
}
\examples{
\dontrun{
  v_get_event(1)
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetEvent.html}
}
