% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media.R
\name{v_get_press_release_list}
\alias{v_get_press_release_list}
\title{Get a list of press releases}
\usage{
v_get_press_release_list(
  no,
  startindex,
  filter,
  fields = v_fields("Press Release")
)
}
\arguments{
\item{no}{integer: the number of press releases to return}

\item{startindex}{integer: index of the first press release to return}

\item{filter}{list: (currently ignored)}

\item{fields}{character: fields to return}
}
\value{
A data.frame
}
\description{
Get a list of press releases
}
\examples{
\dontrun{
  v_get_press_release_list()
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetPressReleaseList.html}
}
