% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referees.R
\name{v_get_referee_list}
\alias{v_get_referee_list}
\title{Get a list of referees}
\usage{
v_get_referee_list(fields = v_fields("Referee"), version, filter)
}
\arguments{
\item{fields}{character: fields to return}

\item{version}{integer: version of local list (currently ignored)}

\item{filter}{list: (currently ignored)}
}
\value{
A data.frame
}
\description{
Get a list of referees
}
\examples{
\dontrun{
  v_get_referee_list()
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#Beach\%20Live_xsd~e-BeachLive~e-Referee.html}
}
