% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{v_get_volley_match}
\alias{v_get_volley_match}
\title{Get a volleyball match}
\usage{
v_get_volley_match(parent, children)
}
\arguments{
\item{parent}{list: List of named charactor vectors. Specify attributes of Request Node.}

\item{children}{list: Named list of named charactor vectors. Add child nodes such as Filter, Relation.}
}
\value{
A data.frame
}
\description{
Get a volleyball match
}
\examples{
\dontrun{
  pl <- list(No = 13754, Fields = paste0(v_fields("Volleyball Match"), collapse = " "))
  v_get_volley_match(parent = pl)
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetVolleyMatch.html}
}
