% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{v_get_volley_player}
\alias{v_get_volley_player}
\title{Get a registration of a player in a volleyball tournament}
\usage{
v_get_volley_player(no, fields)
}
\arguments{
\item{no}{integer: the number of the player registration}

\item{fields}{character: fields to return}
}
\value{
A data.frame
}
\description{
Get a registration of a player in a volleyball tournament
}
\examples{
\dontrun{
  v_get_volley_player(2508)
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetVolleyPlayer.html}
}
