% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{v_get_volley_player_list}
\alias{v_get_volley_player_list}
\title{Get a list of registration of players in volleyball tournaments}
\usage{
v_get_volley_player_list(
  parent = list(Fields = paste0(v_fields("Volleyball Player"), collapse = " ")),
  children
)
}
\arguments{
\item{parent}{list: List of named charactor vectors. Specify attributes of Request Node.}

\item{children}{list: Named list of named charactor vectors. Add child nodes such as Filter, Relation.}
}
\value{
A data.frame
}
\description{
Get a list of registration of players in volleyball tournaments
}
\examples{
\dontrun{
  cl <- list(
    Filter = c(NoTournament = 1257),
    Relation = c(Name = "Player", Fields = "No TeamName FirstName LastName")
  )
  v_get_volley_player_list(children = cl)
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetVolleyPlayerList.html}
}
