% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{v_get_volley_player_ranking}
\alias{v_get_volley_player_ranking}
\title{Get the ranking of the players in a volleyball tournament}
\usage{
v_get_volley_player_ranking(parent)
}
\arguments{
\item{parent}{list: List of named charactor vectors. Specify attributes of Request Node.}
}
\value{
A data.frame
}
\description{
Get the ranking of the players in a volleyball tournament
}
\examples{
\dontrun{
  pl <- list(No = 229, NumberOfScorers = 20, NumberOfPlayers = 20,
             Skills = "Block Dig Libero Reception Scorer Service Set Spike")
  vb_player_ranking <- v_get_volley_player_ranking(parent = pl)
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetVolleyPlayersRankingList.html}
}
