% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pools.R
\name{v_get_volley_pool}
\alias{v_get_volley_pool}
\title{Get a volleyball pool}
\usage{
v_get_volley_pool(parent, children)
}
\arguments{
\item{parent}{list: List of named charactor vectors. Specify attributes of Request Node.}

\item{children}{list: Named list of named charactor vectors. Add child nodes such as Filter, Relation.}
}
\value{
A data.frame
}
\description{
Get a volleyball pool
}
\examples{
\dontrun{
  pl <- list(No = 13754, Fields = paste0(v_fields("Volleyball Pool"), collapse = " "))
  v_get_volley_pool(parent = pl)
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetVolleyPool.html}
}
