% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pools.R
\name{v_get_volley_pool_list}
\alias{v_get_volley_pool_list}
\title{Get a list of volleyball pools}
\usage{
v_get_volley_pool_list(
  parent = list(Fields = paste0(v_fields("Volleyball Pool"), collapse = " ")),
  children
)
}
\arguments{
\item{parent}{list: List of named charactor vectors. Specify attributes of Request Node.}

\item{children}{list: Named list of named charactor vectors. Add child nodes such as Filter, Relation.}
}
\value{
A data.frame
}
\description{
Get a list of volleyball pools
}
\examples{
\dontrun{
  cl <- list(Filter = c(NoTournament = 1257))
  v_get_volley_pool_list(children = cl)
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetVolleyPoolList.html}
}
