% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pools.R
\name{v_get_volley_pool_ranking}
\alias{v_get_volley_pool_ranking}
\title{Get the ranking of a volleyball pool}
\usage{
v_get_volley_pool_ranking(parent)
}
\arguments{
\item{parent}{list: List of named charactor vectors. Specify attributes of Request Node.}
}
\value{
A data.frame
}
\description{
Get the ranking of a volleyball pool
}
\examples{
\dontrun{
  pl <- list(NoPool = "3354", Fields = paste(v_fields("Volleyball Pool Ranking"), collapse = " "))
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetVolleyPoolRanking.html}
}
