% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{v_get_volley_team_ranking}
\alias{v_get_volley_team_ranking}
\title{Get the ranking of the teams in a volleyball tournament}
\usage{
v_get_volley_team_ranking(parent)
}
\arguments{
\item{parent}{list: List of named charactor vectors. Specify attributes of Request Node.}
}
\value{
A data.frame
}
\description{
Get the ranking of the teams in a volleyball tournament
}
\examples{
\dontrun{
  pl <- list(No = 229, NumberOfScorers = 20, NumberOfPlayers = 20, Skills = "Block Dig Libero Reception Scorer Service Set Spike")
  vb_team_ranking <- v_get_volley_team_ranking(parent = pl)
}

}
\references{
\url{https://www.fivb.org/VisSDK/Fivb.Vis.DataWPF/Fivb.Vis.DataWPF~Fivb.Vis.Volley.GetVolleyTeamsRankingListRequest_members.html}
}
