% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournaments.R
\name{v_get_volley_tournament_list}
\alias{v_get_volley_tournament_list}
\title{Get a list of volleyball tournaments}
\usage{
v_get_volley_tournament_list(
  fields = v_fields("Volleyball Tournament"),
  version,
  filter
)
}
\arguments{
\item{fields}{character: fields to return}

\item{version}{integer: version of local list (currently ignored)}

\item{filter}{list: (currently ignored)}
}
\value{
A data.frame
}
\description{
Get a list of volleyball tournaments
}
\examples{
\dontrun{
  v_get_volley_tournament_list()
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetVolleyTournamentList.html}
}
