% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournaments.R
\name{v_get_volley_tournament_ranking}
\alias{v_get_volley_tournament_ranking}
\title{Get the ranking of a volleyball tournament}
\usage{
v_get_volley_tournament_ranking(
  no,
  fields = v_fields("Volleyball Tournament Ranking")
)
}
\arguments{
\item{no}{integer: the tournament number ("No" as returned by \code{v_get_volley_tournament_list}}

\item{fields}{character: fields to return}
}
\value{
A data.frame
}
\description{
Get the ranking of a volleyball tournament
}
\examples{
\dontrun{
  v_get_volley_tournament_ranking(1)
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetVolleyTournamentRanking.html}
}
