% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{v_options}
\alias{v_options}
\title{Set or get fivbvis options}
\usage{
v_options(verbose = FALSE)
}
\arguments{
\item{verbose}{logical: if \code{TRUE}, display progress information}
}
\value{
The current options (invisibly unless at least one parameter has been provided)
}
\description{
Set or get fivbvis options
}
