% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{v_fields}
\alias{v_fields}
\title{Return the fields for a given VIS data type}
\usage{
v_fields(type)
}
\arguments{
\item{type}{string: the data type, currently one of "Volleyball Tournament", "Volleyball Tournament Filter", "Volleyball Tournament Ranking", "Volleyball Match", "Volleyball Match Filter", "Volleyball Pool", "Volleyball Pool Filter", "Volleyball Pool Ranking", "Volleyball Team", "Volleyball Team Filter", "Volleyball Player", "Volleyball Player Filter", "Volleyball Player Ranking", "Volleyball Statistics"}
}
\value{
A character vector of field names
}
\description{
Return the fields for a given VIS data type
}
\examples{
v_fields("Volleyball Tournament")

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/}
}
