% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beach.R
\name{v_get_beach_olympic_selection_ranking}
\alias{v_get_beach_olympic_selection_ranking}
\title{Request to get a beach volleyball Olympic Selection ranking}
\usage{
v_get_beach_olympic_selection_ranking(
  gender,
  gamesyear,
  fields = v_fields("Beach Olympic Selection Ranking")
)
}
\arguments{
\item{gender}{character: gender of the olympic ranking selection}

\item{gamesyear}{integer: year to return ranking selection. if this parameter is not specified, the web service will return the ranking for the latest olympic games with a ranking.}

\item{fields}{character: fields to return}
}
\value{
A data.frame
}
\description{
Request to get a beach volleyball Olympic Selection ranking
}
\examples{
\dontrun{
  v_get_beach_olympic_selection_ranking("W")
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetBeachOlympicSelectionRanking.html}
}
