% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{v_get_event_list}
\alias{v_get_event_list}
\title{Get an event list}
\usage{
v_get_event_list(fields = v_fields("Event"), version, filter)
}
\arguments{
\item{fields}{character: fields to return}

\item{version}{integer: version of local list (currently ignored)}

\item{filter}{list: (currently ignored)}
}
\value{
A data.frame
}
\description{
Get an event list
}
\examples{
\dontrun{
v_get_event_list()
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetEventList.html}
}
