% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media.R
\name{v_get_press_release_text}
\alias{v_get_press_release_text}
\title{Get a press release text}
\usage{
v_get_press_release_text(no, fields = v_fields("Press Release Text"))
}
\arguments{
\item{no}{integer: press release number}

\item{fields}{character: fields to return}

\item{language}{character: (currently ignored)}

\item{filter}{list: (currently ignored)}
}
\value{
A data.frame
}
\description{
Get a press release text
}
\examples{
\dontrun{
  v_get_press_release_text()
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetPressReleaseText.html}
}
