% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{v_get_volley_match}
\alias{v_get_volley_match}
\title{Get a volleyball match}
\usage{
v_get_volley_match(no, fields)
}
\arguments{
\item{no}{integer: match number}

\item{fields}{character: fields to return}
}
\value{
A data.frame
}
\description{
Get a volleyball match
}
\examples{
\dontrun{
  v_get_volley_match(7604)
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetVolleyMatch.html}
}
