% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{v_get_volley_match}
\alias{v_get_volley_match}
\title{Get a volleyball match}
\usage{
v_get_volley_match(no, fields, parent, children)
}
\arguments{
\item{no}{integer: match number. Provide either \code{no} (and optionally \code{fields}) OR \code{parent} (and optionally \code{children}). See Details}

\item{fields}{character: fields to return. Only relevant if \code{no} has been provided}

\item{parent}{list: List of named charactor vectors. Specify attributes of Request Node. Ignored if \code{no} has been provided}

\item{children}{list: Named list of named charactor vectors. Add child nodes such as Filter, Relation. Ignored if \code{no} has been provided}
}
\value{
A data.frame
}
\description{
Get a volleyball match
}
\details{
Parameters can be provided in two ways. The simple interface (retained for backwards compatibility) accepts \code{no} and optionally \code{fields}. More complex queries can be issued using the \code{parent} and \code{children} parameters.
}
\examples{
\dontrun{

  ## using the simple interface
  v_get_volley_match(13754)

  ## or equivalently
  pl <- list(No = 13754, Fields = paste0(v_fields("Volleyball Match"), collapse = " "))
  v_get_volley_match(parent = pl)
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetVolleyMatch.html}
}
