% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{v_get_volley_player}
\alias{v_get_volley_player}
\title{Get a registration of a player in a volleyball tournament}
\usage{
v_get_volley_player(parent, children)
}
\arguments{
\item{parent}{list: List of named charactor vectors. Specify attributes of Request Node.}

\item{children}{list: Named list of named charactor vectors. Add child nodes such as Filter, Relation.}
}
\value{
A data.frame
}
\description{
Get a registration of a player in a volleyball tournament
}
\examples{
\dontrun{
  pl <- list(No = 89538, Fields = paste0(v_fields("Volleyball Player"), collapse = " "))
  cl <- list(Relation = c(Name = "Player", Fields = "No TeamName FirstName LastName"))
  v_get_volley_player(parent = pl, children = cl)
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetVolleyPlayer.html}
}
