% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{v_get_volley_player_list}
\alias{v_get_volley_player_list}
\title{Get a list of registration of players in volleyball tournaments}
\usage{
v_get_volley_player_list(
  fields = v_fields("Volleyball Player"),
  version,
  filter
)
}
\arguments{
\item{fields}{character: fields to return}

\item{version}{integer: version of local list (currently ignored)}

\item{filter}{list: (currently ignored)}
}
\value{
A data.frame
}
\description{
Get a list of registration of players in volleyball tournaments
}
\examples{
\dontrun{
  v_get_volley_player_list()
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetVolleyPlayerList.html}
}
