% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournaments.R
\name{v_get_volley_tournament}
\alias{v_get_volley_tournament}
\title{Get a volleyball tournament}
\usage{
v_get_volley_tournament(no, fields)
}
\arguments{
\item{no}{integer: the tournament number ("No" as returned by \code{v_get_volley_tournament_list}}

\item{fields}{character: fields to return}
}
\value{
A data.frame
}
\description{
Get a volleyball tournament
}
\examples{
\dontrun{
  v_get_volley_tournament(1)
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetVolleyTournament.html}
}
