% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournaments.R
\name{v_get_volley_tournament}
\alias{v_get_volley_tournament}
\title{Get a volleyball tournament}
\usage{
v_get_volley_tournament(parent, children)
}
\arguments{
\item{parent}{list: List of named charactor vectors. Specify attributes of Request Node.}

\item{children}{list: Named list of named charactor vectors. Add child nodes such as Filter, Relation.}
}
\value{
A data.frame
}
\description{
Get a volleyball tournament
}
\examples{
\dontrun{
  pl <- list(No = 1257)
  v_get_volley_tournament(parent = pl)
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetVolleyTournament.html}
}
