% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournaments.R
\name{v_get_volley_tournament_list}
\alias{v_get_volley_tournament_list}
\title{Get a list of volleyball tournaments}
\usage{
v_get_volley_tournament_list(
  parent = list(Fields = paste0(v_fields("Volleyball Tournament"), collapse = " ")),
  children
)
}
\arguments{
\item{parent}{list: List of named charactor vectors. Specify attributes of Request Node.}

\item{children}{list: Named list of named charactor vectors. Add child nodes such as Filter, Relation.}
}
\value{
A data.frame
}
\description{
Get a list of volleyball tournaments
}
\examples{
\dontrun{
  cl <- list(Filter = c(Seasons = "2022"))
  v_get_volley_tournament_list(children = cl)
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetVolleyTournamentList.html}
}
