% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournaments.R
\name{v_get_volley_tournament_ranking}
\alias{v_get_volley_tournament_ranking}
\title{Get the ranking of a volleyball tournament}
\usage{
v_get_volley_tournament_ranking(parent)
}
\arguments{
\item{parent}{list: List of named charactor vectors. Specify attributes of Request Node.}
}
\value{
A data.frame
}
\description{
Get the ranking of a volleyball tournament
}
\examples{
\dontrun{
  pl <- list(
    NoTournament = "1182",
    Fields = paste(v_fields("Volleyball Tournament Ranking"), collapse = " ")
  )
  vb_tournament_ranking <- v_get_volley_tournament_ranking(parent = pl)
}

}
\references{
\url{https://www.fivb.org/VisSDK/VisWebService/#GetVolleyTournamentRanking.html}
}
