% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{v_options}
\alias{v_options}
\title{Set or get fivbvis options}
\usage{
v_options(huge_xml = FALSE, verbose = FALSE)
}
\arguments{
\item{huge_xml}{logical: if \code{TRUE}, set the XML parsing options to allow "huge" (i.e. deeply-nested) XML documents to be parsed}

\item{verbose}{logical: if \code{TRUE}, display progress information}
}
\value{
The current options (invisibly unless at least one parameter has been provided)
}
\description{
Set or get fivbvis options
}
